[REFACTORING] Extract a CassandraACLDAO
This enables separation between the denormalization logic and
the controversial choices made in the implementation of this class (JSON, LWT)
